report telemetry
